Plumb through extra_data provided by diff parsers.
This change makes it so custom diff parsers (associated with specific
SCMTools) can provide additional extra_data to be set on the FileDiff.
This change does the plumbing to take the extra_data from the parser and
store it in the database.

Testing Done:
- Ran unit tests.
- Used in conjunction with other changes.

Reviewed at https://reviews.reviewboard.org/r/11295/