Fix the size of the add/delete marker for changed commit lists.
The change description entry for commit lists had an overly-wide marker
column (showing "-" or "+"), due to relying on the browser's native
sizing of table columns. This is a trivial change that applies the same
`width: 100%` rule we use for change entries for text fields, to ensure
the commit message takes up as much room as possible.