Fix the dashboard size after a periodic reload.
Some work on the JavaScript for the datagrid pages regressed the ability
for the datagrid to be bound to viewport edges. We had code that would
update the size properly, but this was removed as the code was reworked.

This change brings back the resize logic during datagrid setup.

Testing Done:
Added some temporary code for letting me trigger a dashboard reload.
Verified that with this fix, the resulting size was correct in all
pages using the dashboard.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/11314/