Add information on repository plans and fields in the Hosting Service API.
Now that the repository API formally supports creating repositories
backed by hosting services, it's important to communicate the fields
that must be set for each repository.

This change enhances the Hosting Service API to show the repository
plans, consisting of their ID, display name, and list of fields. This
isn't comprehensive. The list of fields doesn't include any type
information, as we'd need to be careful how we represent that from form
fields, but they do include the field name, display name, required flag,
and help text. This should be enough at this stage for API consumers.

Testing Done:
Unit tests passed.

Manually verified the results in the payload.

Built the docs and checked that the payloads contained the information as
well.

Reviewed at https://reviews.reviewboard.org/r/11323/