Remove usage of our custom MySQL backend.
While Review Board 4.0 was early in development, we hit a case with
Django's MySQL backend where setting a `BinaryField` to binary contents
would trigger a warning. This was fixed in Django 1.10.5, but applied to
Django 1.6 as well.

We worked around this in a custom backend defined in Djblets that we
patched in in Review Board's settings. As we're no longer using any
versions of Django that are susceptible to this bug, we no longer need
to patch in the backend.

This change removes that, allowing the original backend to be used.

Testing Done:
Tested an install running on MySQL.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/11342/