Update ClearCase to use force_text instead of force_unicode.
A recent change introduced a `force_unicode` import, which was
deprecated in Django and removed in 1.11. This broke ClearCase on
`release-4.0.x`. The modern import is `force_text`, which this change
switches to.

Testing Done:
Succesfully loaded the ClearCase SCMTool on `release-4.0.x` without
errors.

Reviewed at https://reviews.reviewboard.org/r/11350/