Fix a regression and improve content when sending test e-mails.
Some cleanup work in the e-mail settings page resulted in the wrong
keyword argument being used when sending a test e-mail. We were passing
`email_from` and not `from_email`. We didn't have any unit tests
covering this, and it wasn't caught in development.

This change fixes up the keyword argument and adds a test suite. It also
improves upon the e-mail's subject, body, and error handling. The new
content more explicitly lists the product in the summary, adds a
confirmation that things look good in the body, includes recipient
information in the logs, and an error message in the message shown in
the UI.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/11353/