Rewrite the Repositories admin dashboard widget.
The new widget displays 8 repositories (rather than 3) in a grid. Each
entry contains the name of the repository, the service (SCMTool type and
hosting service), and action links for viewing the other repositories or
adding a new one.

If there are no repositories configured, helpful instructions are added
for configuring a new one, along with a link to the manual.

This new implementation is based on top of the new `BaseAdminWidget`
class, moving away from the old `Widget`. It also uses CSS Grid to lay
out the repositories. While we haven't opted to use this elsewhere in
the UI, this is safe for this limited case.

Testing Done:
Tested with and without repositories, and in desktop and mobile modes.

Reviewed at https://reviews.reviewboard.org/r/11354/