Remove rb-site's global site variable.
This removes the global `site` variable from rb-site, which defines the
site that a command will operate on. Instead of managing this as a
global, it's now passed to `Command.run()`, which can store and operate
on that.

This improves testability and cleans up the code.

Testing Done:
Tested all commands thoroughly.

Reviewed at https://reviews.reviewboard.org/r/11365/