Fix a display issue with the Reviews/Diff tabs on Firefox.
We've had a long-standing issue where the Reviews and Diff tabs weren't
connected quite right to the review request box. They were fine on
Chrome and Safari, but would either appear to be connected with a solid
line or a fairly thin but visible line to the rest of the box on
Firefox.

This ended up being (at least on a high-DPI display) an issue with a
half-pixel height difference between these tabs and the rest of the
container. That came from the Star/Archive icons on the left-hand side
of the review request action bar, or, more specifically, their
`line-height`. We set that to inherit, but the default line height ended
up doing the wrong thing in our case.

We now set this explicitly to 0. This does not impact the display at
all, and it fixes the alignment issue.

Testing Done:
Verified the display on Firefox. Also verified there weren't regressions
on Safari, Chrome, or Edge.

Reviewed at https://reviews.reviewboard.org/r/11371/