Use Djblets to build the Review Board pipeline configuration.
Our Pipeline configuration is pretty complex, and we duplicate a lot of
specific settings from Djblets. Now that Djblets has a function for
handling the generation of this standard configuration, we can use it to
build the one in Review Board. This simplifies settings quite a bit and
makes static media compilation and packaging more resilient to
dependency or plugin changes in Djblets.

Testing Done:
Successfully collected static media and built a package.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/11368/