Rework the support collection settings and instructions for rb-site.
This introduces an `--opt-in-support-data` flag for `rb-site install`
that turns on support data collection. The old flag,
`--opt-out-support-data`, is now default. These flags are used only for
automated installs.

When prompting for this question, the user must give an explicit yes or
no answer. To help make this choice, this also expands the information
about what we collect and how it will be used.

Testing Done:
Ran `rb-site install` and verified the resulting content was correct.

Verified the opt in/out flags set the right state.

Reviewed at https://reviews.reviewboard.org/r/11374/