Allow a secret key to be provided with rb-site install --secret-key.
A common issue when running `rb-site install` in an automated setup (for
Docker instances or when using a tool like Chef) is that each new
install would get its own `SECRET_KEY` setting, requiring users to
override the resulting file.

To address this, a new `--secret-key` option has been added, which
allows the caller to provide an explicit key to use. This must be 50
characters or higher (matching what we generate ourselves). If this is
set, it will be used instead of generating a new key.

Testing Done:
Tested creating a new site with and without `--secret-key`.

Tested the character length validation.

Reviewed at https://reviews.reviewboard.org/r/11375/