Add support for custom settings_local.py templates in rb-site.
We've had a long-standing request from a customer to allow the
`settings_local.py` template to be customizable via a template. This
change enables this through a `--settings-local-template` parameter,
which can point to a local file that will be processed and turned into
the resulting `settings_local.py`.

The help output displays the default template file we use, so that
consumers can better find something to base theirs off of.

Other templates are not (at this time) customizable, though we could now
more easily do this in the future. Most of our templates are for web
servers or cron, rather than part of the Review Board runtime, and are
therefore less critical for the site installation.

Testing Done:
Tested with and without a custom template. Saw the expected result in
each case.

Reviewed at https://reviews.reviewboard.org/r/11376/