Fix encoding issues with LDAP/AD on Python 3.x
A user reported encoding errors in their logs when attempting to log in
a new user when connected to LDAP. The issue here was that data returned
from ldap was `bytes`, but we were attempting to split it using a
`unicode` splitter. According to the ldap documentation, field names are
returned as the text type, but attribute values will be bytestrings. The
documentation is a unclear in that it also claims that "text values"
will be the unicode type. In order to accomodate potentially different
behavior across Python versions, all values have been wrapped in
`force_text`. The unit tests have been updated to supply the expected
(bytes) types for values.

Testing Done:
Ran unit tests on Python 2.7 and 3.8.

Reviewed at https://reviews.reviewboard.org/r/11424/