Document and clean up the UI and ConsoleUI code in rb-site.
When `rb-site` was first written, back in pre-1.0, it supported a
console UI and a GTK+ UI, to help get set up on graphical Linux
installs. We've long since removed the graphical UI, but left the base
foundation of it around, with no peer.

This change strips out the base class, leaving only `ConsoleUI` around.
In the process, it documents it and improves a little of the
functionality to help with reuse.

This is a precursor to moving this file and splitting up some of the
responsibilities so that some of the functionality can be used by
`rbext`, and so `rb-site` can be simplified a bit.

Testing Done:
Tested creation and upgrades of sites. All functionality worked as
expected.

Reviewed at https://reviews.reviewboard.org/r/11407/