Update the dependencies that rb-site checks, and install instructions.
`rb-site` had a fairly outdated list of dependencies to check. Some of
these didn't matter and never showed up in the instructions (`sqlite3`
support, which is part of Python, and `python-memcached`, which we have
as a required dependency). A couple others are still relevant, but the
original packages providing the module have since been forked and
renamed.

This change trims down the module list to what's relevant today, and
enhances them with package names so that we can tell the user what to
run in order to get the support they need.

Testing Done:
Ran through `rb-site install`, with missing modules. Saw the instructions
for installing them.

Reviewed at https://reviews.reviewboard.org/r/11416/