Warn users that extensions aren't loaded until an upgrade is performed.
Review Board avoids loading extensions until the site has been upgraded,
in order to ensure there aren't any breakages. This can be confusing if
one is unaware that the site wasn't upgraded and are trying to use a
management command provided by an extension. While unlikely in
production, it's not uncommon in development.

Now, when this situation is it, a message will be logged informing the
user that extensions won't be loaded until the site is upgraded.

Testing Done:
Tested trying to run an extension-provided management command on a database
not yet upgraded. The command failed, of course, but I saw the warning.

Reviewed at https://reviews.reviewboard.org/r/11444/