Update the documentation on installing development versions.
We have a very old set of docs on how to install development versions of
Review Board. This was out-of-date, and though it wasn't directly linked
to in the main index pages of the manual, it was still available.

This change updates these docs to be more future-proof and useful to
link to, and directs users to the release notes of a particular release
for more specific instructions.

Testing Done:
Viewed the docs. Tested links, build output, and spelling.

Reviewed at https://reviews.reviewboard.org/r/11472/