Fix showing the base directory on the New Review Request page.
A typo in a context variable name led to a regression with uploading
diffs in the New Review Request page that required a base directory. The
correct variable name was `requiresBasedir`. but it was typoed to
`requiresBaseDir`.

This change fixes the typo, and adds unit tests to ensure this and other
context variables won't regress in the future.

Testing Done:
Unit tests pass.

Verified the field was present for Subversion repositories.

Bugs closed: 4911

Reviewed at https://reviews.reviewboard.org/r/11470/