Pin bcrypt to a version compatible with Python 2.7.
`bcrypt` 3.2 dropped support for Python 2.7. This wasn't an issue if
installing Review Board via `pip`, but if using `easy_install`, the
wrong version would be picked up.

The 3.1.x series still supports Python 2.7, so we're now pinning that.

We don't directly depend on `bcrypt`, but `cryptography` and `paramiko`
do. When we drop Python 2.7 support, we can remove this dependency
completely.

Testing Done:
Tested an installation on Python 2.7, and verified that `bcrypt` 3.1.7
was installed.

Bugs closed: 4913

Reviewed at https://reviews.reviewboard.org/r/11471/